separate component for worksheet list item and disable hover interaction if >30 courses
